Move utility fields under `utils` module
